Dont raise tileerror for cancelled map requests
tileerror occurs for a failed GBIF API request.

Previously if a user made several rapid requests, then any in progress requests (for a previous map state)
would be cancelled and this would raise a tileerror
This is not an error, so now behaviour is just to check on load for a tileerror.

This initial check instead guards against a bad taxon ID and/or the GBIF API being down.

(see discussion in PR #98 for details).