Rename trackType config.yml key to displayType
It turns out that in JBrowse, a track's type specifically refers to
FeatureTrack, QuantiativeTrack, etc. To avoid confusion, the name is
changed to better explain that it is the display type (or display model,
the JBrowse docs is not that stringently worded for this attribute) that
we want to control with this key.