Update form_parser to handle doxc v1.2
A big change to the form is that the new layout renders as markdown
tables. This is trickier to parse than the plain text of v1.1. But the
docx version of the form does look nicer for the users. Helper functions
are needed to strip markdown tables. The most complex parsing is for
multi-field multi-line entries like Species name and Species Image
Attribution.